penpals-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
chromadb
requests==2.31.0
python-dotenv==1.0.0
minio
sqlalchemy>=2.0.35